+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.64.0] - 2023-10-02
+### Added
+* Validation to shop related files. Shop files that need to be loaded in a specific order, needs to be accompanied by a
+cog shop config. Validation will require this of the user.
+
 ## [0.63.0] - 2023-10-02
 ### Added
 * Field `source` to `Scenario` and `ModelTemplate` in `CogSHOP` model.